Use an IP and add carry, also jump
So, now, instead of just `pop`ping instructions, we have an instruction
pointer we can play with. That's gonna be useful.

An example of where it's useful is the jump instruction, just now
implemented.
Just has two modes: "always" and "carry".

"always" will... always jump to that IP.
"carry" will only jump to that IP if the carry is true (and everytime
reset the carry).

You can, for now, only change the carry with "inverse_carry".
Obviously, in the future, the VM will support comparisons, but since,
for now, we don't have variables or the like, it's not as interesting.